Refactor: rename and seperate api
