+++ src/api/game.js
@@ -1,38 +1,5 @@
 const API_URL = process.env.REACT_APP_API_ADDRESS;
 
-export const fetchUser = async (tokenId, path) => {
-  try {
-    const res = await fetch(`${API_URL}/users/${path}`, {
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${tokenId}`,
-      },
-      credentials: "include",
-    });
-    const { result, data: user, isNewUser } = await res.json();
-
-    return { result, user, isNewUser };
-  } catch (err) {
-    console.error(err);
-  }
-};
-
-export const deleteUser = async () => {
-  try {
-    const res = await fetch(`${API_URL}/users/logout`, {
-      headers: {
-        "Content-Type": "application/json",
-      },
-      credentials: "include",
-    });
-    const { result } = await res.json();
-
-    return result;
-  } catch (err) {
-    console.error(err);
-  }
-};
-
 export const fetchSchedule = async () => {
   try {
     // TO DO: 실서버 연결시 주석 해제 및 testDay를 today로 교체
@@ -73,6 +40,14 @@ export const fetchPlayers = async () => {
     const { data } = await res.json();
 
     return data;
+  } catch (err) {
+    console.error(err);
+  }
+};
+
+export const postBetting = async (bettingData) => {
+  try {
+    // const res = await fetch(`${API_URL}/games/${testDay}/betting`);
   } catch (err) {
     console.error(err);
   }
